#93 Migrate AttemptCount class
It now resides in the sampling.hpp header instead of its own. This
reduces the number of files we have in the code base.
